test: move env -S shebang tests into cli/install/shebang-normalize.test.ts
`test/regression/issue/` is reserved for true regressions (worked in a
previous release, then broke). `env -S` handling never worked in any
prior release — the Zig spec was already wrong — so this is a
never-worked bug, not a regression. Fold the tests into the existing
file that already covers bin-linking shebang behavior.
